+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 - added Tests for NET Module
 - added Config File in `config_dir()/@shadowdara/flua/config.lua`
 - loading the Config File
+- added config file Infos to the help message
 
 ## 0.2.0